Update tools + Synth Elaborate Only
