+++ openlane/__version__.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-__version__ = "2.0.0a46"
+__version__ = "2.0.0a47"
 
 if __name__ == "__main__":
     print(__version__, end="")
